Rename cross module to compat
