+++ webstruct/compat.py
